Refactor verifier into verify command
