+++ presto-verifier/pom.xml
@@ -54,6 +54,11 @@
 
         <dependency>
             <groupId>io.airlift</groupId>
+            <artifactId>airline</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>io.airlift</groupId>
             <artifactId>bootstrap</artifactId>
         </dependency>
